+++ src/containers/Discover/Discover.tsx
@@ -1,24 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import './Discover.css';
-import gql from 'graphql-tag';
-import { useQuery } from '@apollo/react-hooks';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSearch } from '@fortawesome/free-solid-svg-icons';
 import Navbar from '../../components/Navbar/Navbar';
-import Linkbar from '../../components/Linkbar/Linkbar';
 import RoadmapList from '../../components/RoadmapList/RoadmapList';
+import Linkbar from '../../components/Linkbar/Linkbar';
 import categories from '../../categories';
 
-
-const GET_ROADMAPS_CATEGORY = gql`
-query roadmaps($category: String, $title: String) {
-  roadmaps(category: $category, title: $title) {
-    id
-    title
-    category
-  }
-}
-`;
 
 interface IRoadmap {
   title: string;
@@ -32,39 +20,12 @@ const Discover: React.FC = () => {
   const [searchInput, setSearchInput] = useState('');
   const [currCategory, setCurrCategory] = useState('');
 
-  // fetching roadmaps from database
-  const { data, loading, refetch } = useQuery(GET_ROADMAPS_CATEGORY);
-
-  // filter for clicked category only
-  const renderCategories = (clickedCat: string) => {
-    setCurrCategory(clickedCat);
-    if (clickedCat === 'Popular') {
-      refetch({ category: '' });
-    } else {
-      refetch({ category: clickedCat });
-    }
-  };
-
-  const renderSearchResults = () => {
-    if (currCategory === 'Popular' || currCategory === '') {
-      refetch({ title: searchInput });
-    } else {
-      refetch({ title: searchInput, category: currCategory });
-    }
-  };
-
-  // on click render roadmaps of this category
+  // on click set state of selected category
   const handleClick = (clicked: string) => {
-    renderCategories(clicked);
+    setCurrCategory(clicked);
     setSearchInput('');
   };
 
-  // render when user types in searchbar only
-  useEffect(() => {
-    renderSearchResults();
-  }, [searchInput]);
-
-  if (loading) return null;
   return (
     <>
       <Navbar />
@@ -84,7 +45,7 @@ const Discover: React.FC = () => {
           </div>
         </label>
       </div>
-      <RoadmapList data={data.roadmaps} />
+      <RoadmapList searchInput={searchInput} currCategory={currCategory} />
     </>
   );
 };
